requirements.txt
accelerate==1.10.0
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cmake==4.1.0
colorama==0.4.6
diskcache==5.6.3
et_xmlfile==2.0.0
faiss-cpu==1.12.0
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.8
httptools==0.6.4
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
llama_cpp_python==0.3.16
lxml==6.0.1
Markdown==3.8.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
psutil==7.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
dotenv==0.9.9
